+++ src/hooks/use-delete-actions.ts
@@ -30,8 +30,8 @@ const generateAppointmentDeletedSnackbar = (
 	if (res.type.includes('fulfilled')) {
 		let snackbarLabel =
 			undoAction === undefined
-				? t('message.snackbar.appt_moved_to_trash', 'Appointment moved to trash')
-				: t('message.snackbar.appointment_permanently_deleted', 'Appointment permanently deleted');
+				? t('message.snackbar.appointment_permanently_deleted', 'Appointment permanently deleted')
+				: t('message.snackbar.appt_moved_to_trash', 'Appointment moved to trash');
 		if (isRecurrentSeries) {
 			snackbarLabel = t(
 				'message.snackbar.series_deleted',
@@ -87,7 +87,7 @@ const generateAppointmentRestoredSnackbar = (
 };
 
 type AccountContext = {
-	isInstance?: boolean;
+	isSingleInstance?: boolean;
 	dispatch: Dispatch;
 	replaceHistory: (a: string) => void;
 	onClose: () => void;
@@ -143,7 +143,7 @@ export const useDeleteActions = (
 			const ctxt = {
 				dispatch,
 				t,
-				isInstance: context.isInstance,
+				isInstance: context.isSingleInstance,
 				createSnackbar,
 				newMessage: newMessage?.text?.[0]
 			};
@@ -152,14 +152,14 @@ export const useDeleteActions = (
 					onBoardClose && onBoardClose();
 					generateAppointmentDeletedSnackbar(res, t, createSnackbar, restoreAppointment);
 				})
-				.then(
+				.then(() => {
 					setTimeout(() => {
 						if (notifyOrganizer && !isCanceled) {
 							onBoardClose && onBoardClose();
 							sendResponse(event, ctxt);
 						}
-					}, 5000)
-				);
+					}, 5000);
+				});
 		},
 		[event, context, createSnackbar, dispatch, notifyOrganizer, t]
 	);
@@ -184,11 +184,12 @@ export const useDeleteActions = (
 			const ctxt = {
 				dispatch,
 				t,
-				isInstance: context.isInstance,
+				isInstance: context.isSingleInstance,
 				createSnackbar,
 				newMessage: newMessage?.text?.[0]
 			};
-			const untilDate = moment(event?.resource?.ridZ).subtract(1, 'day').format('YYYYMMDD');
+			const eventDate = event?.resource?.ridZ ?? event.start.valueOf();
+			const untilDate = moment(eventDate).subtract(1, 'day').format('YYYYMMDD');
 			const deleteFunction = (): void => {
 				const modifiedInvite = {
 					...invite,
@@ -216,7 +217,7 @@ export const useDeleteActions = (
 					invite: modifiedInvite,
 					context: {
 						dispatch: context.dispatch,
-						isInstance: context.isInstance,
+						isInstance: context.isSingleInstance,
 						folders: context.folders
 					}
 				});
@@ -255,7 +256,7 @@ export const useDeleteActions = (
 				dispatch,
 				newMessage: newMessage?.text?.[0],
 				t,
-				isInstance: context.isInstance,
+				isInstance: context.isSingleInstance,
 				createSnackbar,
 				inst: {
 					d: invite?.start?.tz