+++ src/test/integration-tests/create-appointment.test.tsx
@@ -13,6 +13,7 @@ import {
 	getMockedAccountItem
 } from '../../carbonio-ui-commons/test/mocks/accounts/fakeAccounts';
 import * as shell from '../../carbonio-ui-commons/test/mocks/carbonio-shell-ui';
+import defaultSettings from '../../carbonio-ui-commons/test/mocks/settings/default-settings';
 import { setupHook, setupTest } from '../../carbonio-ui-commons/test/test-setup';
 import { PREFS_DEFAULTS } from '../../constants';
 import { useOnClickNewButton } from '../../hooks/on-click-new-button';
@@ -20,10 +21,12 @@ import { reducers } from '../../store/redux';
 import BoardEditPanel from '../../view/editor/editor-board-wrapper';
 import mockedData from '../generators';
 
-jest.setTimeout(20000);
+jest.setTimeout(30000);
 
 shell.getUserSettings.mockImplementation(() => ({
+	...defaultSettings,
 	prefs: {
+		...defaultSettings.prefs,
 		zimbraPrefUseTimeZoneListInCalendar: 'TRUE',
 		zimbraPrefTimeZoneId: 'Europe/Berlin',
 		zimbraPrefCalendarDefaultApptDuration: '60m',